Just bundle new relic gems not yaml.
As this file is created by Puppet for each environment, we need to test
that it creates one for prod. Even though the cms is not deployed
through puppet, there could be some config in there which initially set
things up. If not then we’ll recommit this yaml file. (+1 squashed commit)
Squashed commits:
[1a0693b] Set up New Relic