6979 Content hub page (CMS)
Previously, our javascript only had functionality to show the correct
form when the user selected it from the dropdown menu.

However, when a user created a new page, or clicked on a page they had
previously created, they would see both forms, this adds the
functionality to detect which option in the dropdown is selected, and
shows the correct form.

Still needs some refactoring, but a nice step forwards