Create fincap seed data for development
This is the start to add fincap data for development purposes.

Seedback gem was added to differentiate between cms data and fincap data
when running seeds.

The current CMS seed rake task does not need any extra work or change on
the README since it can be run by rake db:seed but only fincap.

With fincap you can run rake db:seed:fincap as stated on the README.

* Moved db/seeds.rb to db/seeds/cms.seeds.rb
* Created db/seeds/fincap.seeds.rb which contain two insight pages
to start with.

Observation: this rake task does not do the markdown convertion to html
on some insight fields yet, because it needs a definition of what doing with
those.