Add blocks_attributes to form builder
This is a refactor. Passing the object through form the controller into
the builder means that it is a lot less hacky to access the current_value

The Comfy Form builder was also causing more issues than help and since
it inherited from ActionView::Helpers::FormBuilder, the decision was made
to inherit directly from there instead.