Index categories and pages for site search
The code uses the Adapter pattern so this can be flexible to change
the indexer service whenever the business needs to.

So for example we have two adapters:

* Local (prints the objects to be indexed - a very basic implementation).
* Algolia (send all the objects to be indexed in the Algolia service).

In order to create more adapters you need to create an object
that responds_to #index_name and #create_or_update and setup
the environment variable INDEXERS_ADAPTER to be the adapter you need.

For the categories we need to separate in welsh and english for the same
record since the category record on database in single for two locales.

For the pages there are still discussions going on about how to
index the content of the pages. There is no conclusion about yet though.

At the moment the Algolia basic account does not support long content
so the implementation decided until this point to use the meta
description field as the content of the page to be indexed.

The index task will be run every morning at 5 a.m.