Add feature tests for indexing cms database
This tests if the objects we are sending through are the right data.

I need to deleted spring from rake because spring was not working
with Aruba gem. That was the only way I saw to solve the issue.
And I don't think there is anyone using Spring on CMS either.