Add simple component to the CMS
This simple component will accept any content basically
and the client will decide how to render those components.
The reason about the name "simple_" is because of the client
will decide how to render the components and not CMS itself.

Also this commit address the fixes for carriage returns on textarea tags