Allow block collection to return all values
Before when we had the following blocks:

<Comfy::Cms::Block id: 13, identifier: "topics", content: "Saving",
  blockable_id: 2, blockable_type: "Comfy::Cms::Page">
<Comfy::Cms::Block id: 14, identifier: "topics", content: "Financial
  Capability", blockable_id: 2, blockable_type: "Comfy::Cms::Page">

Then the API were return two topics but with "Saving" as value.
This was returning duplicated values because the Blcok retriever
class was using Array#find.

After reading the class I realised that the blocks retriever only exists
to retrieve the live page body content and from nothing else,
so then it makes sense for me to "use block retriever for content
and nothing from the blocks".