Add domain specific environment variables to switch between apps
Because we are using this same source code to run multiple CMS applications
on the same server each application has to know some specifics to ensure they
do not interfere with each other. Environment variables are passed to each application
on boot and the app can then use these. Most of this logic currently sits in the new
Domain.rb. There is potentially some further refactoring that can be done for this logic
once we have upgraded to Rails 4.2, but until now this is perhaps a clearer way of defining
these variables.

There is also some further improvements that can be made for developers so they do not have to
remember or think as much when working locally. particularly around the seeds etc. Two separate
databases can be created and seeded separately. This has yet to be implemented, but there is documentation
in the README as to how to keep the environments separate when working locally