Rewrite the tag converter
Nokogiri was gaving to much headache to make work with document fragments.
Mainly because I need to change on every format the text to add
html tags and nokogiri was not playing right for manipulating the same text
adding tags and finding the index on a string.
So I decided to use pure ruby instead.