Add convert to html rake prismic task
I created this task so the Q.As can test the tranformation from Prismic
into pure HTML and markdown.

This task will generate all prismic document into the tmp folder.
E.g. /tmp/a-document-from-prismic.html

On each of html I did by this format:
Attribute "the name of the attribute". And then the value converted html.