Remove markdown editor strip functionality on init
This undocumented piece of code was defined as 'hack' to some unknown issue
(https://github.com/moneyadviceservice/cms/commit/065437316e459291bd57578baefa9acba1128f25)

This automatic trim would be applied only when the editor is being loaded with some existing content,
it's not an automated trim on the markdown as-you-type or at save time.

Therefore, it's probably uneeded and it's causing issues with indented content (i.e. nested lists).

Co-Authored-By: Tomas D'Stefano <tomas_stefano@successoft.com>