Add `order_by_date` param to documents EP
The documents EP should take an optional param in order to override the
default order (by `position`, followed by `created_at` for the average case).

Due to the nature of the ordering (needs to look into the `content` and parse
a potential date), it is not compatible with every other chained scope of
`@documents` in the `DocumentProvider`, but it's been tested to work
with `news` as `document_type` and no other filters applied.

Since we are not planning to use this ordering for anything other than
the latest news, this should be a safe addition.