[DEV APPROVED] - [9488] Teaser section title (#487)
* Add rake task to reset layouts and associated blocks

* Replace _ with - to avoid wrong slug being generated

* Add blocks count before updating the database

Calling the count after the update makes active record do another
call to the db. Since is after the update the information is
displayed "Updated 0 records" which is not true. Adding before
makes active record do the count before update and display
the right message: "Updating 9 records".
Also add a puts '=' * 120 just to idenfy the text in the midst
of SQL queries.