Add a new What Works page layout
- Amend fincap:setup rake task. This will provide a new what_works page layout
  and provide a new URL scope. i.e /en/what_works/my-custom-page
- Add what_works seed in order to populate the DB with What Works data
- Add "what_works" specific inflection rule

The default Rails infector was transforming "what_works" into
"what_work".
This, combined with the fact that our Content Controller singularizes
the page types before looking for them in the list of existing
layouts, was causing that the CMS Client requests from Fincap for
"what_works" pages were being mapped into looking for a "what_work"
layout (that doesn't exist) and then erroring with a 400 code.