Isolate Fincap Layout definitions into single file
During the setup of a new Fincap layout we realised that we were
pasting the same block of content in three different places:
- Seeds File
- Fincap rake task
- Feature test

Having to paste the same block of content in those places every time
we add a new layout seems to me annoying and prone to errors.
Due to that, this is extracting the new layouts setup into a single
config file, and the same configuration can be used later in multiple
places without having to paste the same content around.