Add task fixing evidence summaries creation date
When the evidence summary pages were migrated to the new FINCAP site,
the 'created_at' field for all of them was set to the date/time of
the migration.

Due to this, when evidence summaries are filtered in the evidence hub,
some really old ones get displayed on top, as the results are ordered
by creation date and these old summaries contain a recent one (the
migration date).

To avoid this, has been decided between tech and product to override
the 'created_at' value of every evindence summary page with year of
publication prior to 2018.
For all the affected pages, the new 'created_at' value will be the
first second of their year of publication.

Eg: 'year_of_publication' is 2016
    'created_at' will be '2016-01-01 00:00:0'