Remove task fixing evidence summaries creation
This was an one time taks that was successfully executed.
There is no plan to execute it again, as the wrong creation timestamps
were caused by the migration of the pages to the new platform
setting a created_at value different from the original information.

Also the task tests are using the DB, so introduce some speed pain for
code that is not mean to be running anymore.