Remove test dependency on specific tag elems order
The test was involuntary asserting that the first element in the
presenter generated input tag must be the "checked" option.

This was the case with Rails 4.1, but once moved to Rails 4.2, when the
generated elements in the input tag were exactly the same, the order of
them was different, not being the "checked" option the first element in
the input tag.

This modification checks for the presence of the same elements but
without asserting on their position.