Adapt file controller tests for presented objects
FilesController#index has an after_action hook presenting the resulting
files.
Previously, with Rails 4.1 & Rspec 3.0, the resulting files got by the
tests were the original files before being presented with an
after_action hook.
With Rails 4.2 & Rspec 3.8, the resulting files are the ones presented
at the after_action hook.

Due to this, as the tests were failing after the Rails update and we
want to check that the files assigned are the ones contained in the
site, we need to modify the tests to explicitely address that we are
assigning presented objects, and that those objects are the site files.

I was tempted to jump into the files controller and start removing
instance variables assignations from the controller hooks to explicitely
call them in the corresponding actions.

There are som caveats when trying to do this:
- We have controller actions inheritance redefinitions that get into
  trouble when the parent already presents the objects instead of using
  a hook around the output of the child&parent combinated result.
- This controller is overriding Comfortable Mexican Sofa code. I would
  prefer not to modify code here unless extrictly necessary as an
  interface change could have unexpected side consequences.